L55: PubSub
Adds a PubSub module for managing real-time updates to clients.

Uses `Phoenix.PubSub` to broadcast raffle and ticket updates.
This provides a mechanism for live updates in the UI,
allowing users to see new tickets as they are created
or raffle details as they are updated without refreshing the page.